♻️ [Refactoring]: #17 4 つの resolve-*.ts を AttrResolver companion object に集約
resolveMediaBox / resolveCropBox / resolveRotate / resolveResources を
それぞれ別ファイルに分けるとファイル数が増えすぎるため、attr-resolver.ts に
companion object として統合する。呼び出し側は AttrResolver.mediaBox /
.cropBox / .rotate / .resources に書き換え（DictReader / IndirectRef と整合）。
テストも attr-resolver.behavior.test.ts 1 ファイルに集約。
